fix(warrior): restore Arms execute, core loop, and 8.3 talents
Carry rage cost onto child casts so Execute, Tactician, and Anger
Management share one payment. Fix auto-attack rage rounding and
bind missing 8.3 scripts. Pending in-game acceptance.

+++ src/server/game/Entities/Unit/Unit.h
@@ -2060,7 +2060,7 @@ class TC_GAME_API Unit : public WorldObject
         bool IsFalling() const;
         virtual bool CanSwim() const;
 
-        void RewardRage(uint32 baseRage);
+        void RewardRage(float baseRage);
 
         virtual float GetFollowAngle() const { return static_cast<float>(M_PI/2); }
